Move abstract-component-action-delegate to ui/abstract
