+++ ui/controls/inspector-footer.reel/inspector-footer.js
@@ -1,7 +1,7 @@
 /**
  * @module ui/inspector-footer.reel
  */
-var AbstractComponentActionDelegate = require("core/ui/abstract-component-action-delegate").AbstractComponentActionDelegate;
+var AbstractComponentActionDelegate = require("ui/abstract/abstract-component-action-delegate").AbstractComponentActionDelegate;
 
 /**
  * @class InspectorFooter
