+++ ui/controls/topology.reel/topology.js
@@ -3,7 +3,7 @@
  */
 var Component = require("montage/ui/component").Component,
     Model = require("core/model/model").Model,
-    AbstractComponentActionDelegate = require("core/ui/abstract-component-action-delegate").AbstractComponentActionDelegate;
+    AbstractComponentActionDelegate = require("ui/abstract/abstract-component-action-delegate").AbstractComponentActionDelegate;
 
 
 var CLASS_NAMES_FOR_MODES = {
