+++ ui/controls/viewer.reel/viewer.js
@@ -1,4 +1,4 @@
-var AbstractComponentActionDelegate = require("core/ui/abstract-component-action-delegate").AbstractComponentActionDelegate,
+var AbstractComponentActionDelegate = require("ui/abstract/abstract-component-action-delegate").AbstractComponentActionDelegate,
     CascadingList = require("ui/controls/cascading-list.reel").CascadingList;
 
 /**
