+++ ui/dashboard/dashboard.reel/dashboard.js
@@ -1,4 +1,4 @@
-var AbstractComponentActionDelegate = require("core/ui/abstract-component-action-delegate").AbstractComponentActionDelegate;
+var AbstractComponentActionDelegate = require("ui/abstract/abstract-component-action-delegate").AbstractComponentActionDelegate;
 
 /**
  * @class Dashboard
