+++ ui/dashboard/notifications.reel/notifications.js
@@ -1,4 +1,4 @@
-var AbstractComponentActionDelegate = require("core/ui/abstract-component-action-delegate").AbstractComponentActionDelegate,
+var AbstractComponentActionDelegate = require("ui/abstract/abstract-component-action-delegate").AbstractComponentActionDelegate,
     Notification = require("ui/dashboard/notifications.reel/notification.reel").Notification,
     notificationCenter = require("core/backend/notification-center").defaultNotificationCenter;
