+++ ui/widgets/alerts.reel/alerts.js
@@ -1,7 +1,7 @@
 /**
  * @module ui/alerts.reel
  */
-var AbstractComponentActionDelegate = require("core/ui/abstract-component-action-delegate").AbstractComponentActionDelegate;
+var AbstractComponentActionDelegate = require("ui/abstract/abstract-component-action-delegate").AbstractComponentActionDelegate;
 
 /**
  * @class Alerts
